[unreal]重构，把ts库拷贝的逻辑挪到单独文件
